Starting to change to using the REST API to obtain the data.
App uses the wordpress-rest-api-oauth-1 package to communicate with the server
which is currently hardcoded as http://qw/bigram

The App class implements the methods that will obtain the data and manage the state.
As we're migrating from the very simple data model to the REST API model
we're using new classes. The new class names are simply given a suffix of 2.

Caveats: Files are a bit of a mess at present.
I felt I wanted to keep some of the original code in App, Board, Sides etc
but copied the code to the files suffixed with 2 a bit too late in the process.
Never mind. If we really want to we can checkout the code at a particular commit point
and make a branch or something.

All the code does at present is to load the s-letter and b-letter taxonomies.
Not very well, since there were problems with per_page requests
and not setting key= parameters properly.